feat(db): Enable eager loading for user group association
Changes:
- Modified: src/typeORM/entities/user.ts

Context:
This commit enables eager loading for the 'group' association in the
User entity. By setting `eager: true`, the user's group will now be
loaded automatically when the user is accessed, reducing the need for
separate queries. This change enhances the efficiency of retrieving user
information along with associated group details.